+++ lib/Conversion/TritonGPUToLLVM/CMakeLists.txt
@@ -1,6 +1,13 @@
 add_mlir_conversion_library(TritonGPUToLLVM
+    TritonGPUToLLVM.cpp
     TritonGPUToLLVMPass.cpp
     PTXAsmFormat.cpp
+    ConvertLayoutOpToLLVM.cpp
+    ElementwiseOpToLLVM.cpp
+    ViewOpToLLVM.cpp
+    LoadStoreOpToLLVM.cpp
+    DotOpToLLVM.cpp
+    ReduceOpToLLVM.cpp
 
     ADDITIONAL_HEADER_DIRS
     ${PROJECT_SOURCE_DIR}/include/triton/Conversion/TritonGPUToLLVM
